+++ src/DependencyInjection/Compiler/NewsAlertPass.php
@@ -21,11 +21,23 @@ class NewsAlertPass implements CompilerPassInterface
      */
     public function process(ContainerBuilder $container)
     {
-        if (!$container->has('hh.contao-newsalert.newstopiccollection')) {
+        if (!$container->has('huh.newsalert.topiccollection')) {
             return;
         }
-        $definition = $container->findDefinition('hh.contao-newsalert.newstopiccollection');
+        $definition = $container->findDefinition('huh.newsalert.topiccollection');
+        
+        $taggedServices = $container->findTaggedServiceIds('huh.newsalert.topic_source');
+
+        foreach ($taggedServices as $id => $tags) {
+            $definition->addMethodCall('addTopicSource', [new Reference($id)]);
+        }
+
         $taggedServices = $container->findTaggedServiceIds('hh.newsalert.topic_source');
+
+        if (count($taggedServices) > 0)
+        {
+            trigger_error("Tag hh.newsalert.topic_source has beed marked deprecated and will be removed in next major newsalert version. Use huh.newsalert.topiccollection instead.", E_USER_NOTICE);
+        }
 
         foreach ($taggedServices as $id => $tags) {
             $definition->addMethodCall('addTopicSource', [new Reference($id)]);