+++ src/Resources/contao/dca/tl_news.php
@@ -14,7 +14,7 @@ $translator = System::getContainer()->get('translator');
 /*
  * Callbacks
  */
-$dc['config']['onsubmit_callback'][] = ['hh.contao-newsalert.listener.newspostedlistener','onSubmitCallback'];
+$dc['config']['onsubmit_callback'][] = ['huh.newsalert.listener.newsposted','onSubmitCallback'];
 
 /*
  * Palettes